remove unnecessary code and clean things up
